# src/backend/api_spec.yaml
openapi: 3.0.1
info:
  title: Sayyara API
  description: API Explorer for Sayyara
  termsOfService: https://localhost/
  contact:
    email: example@example.com
  license:
    name: Closed License
  version: v1
servers:
  - url: http://localhost/api
security:
  - Bearer: []
paths:
...
  /shop/{id}:
    get:
      tags:
        - Shop
      operationId: shop_read
      parameters:
        - name: id
          in: path
          description: A unique integer value identifying this shop.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
    put:
      tags:
        - Shop
      description: Update a given shop
      operationId: shop_update
      parameters:
        - name: id
          in: path
          description: A unique integer value identifying this shop.
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Shop'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
      x-codegen-request-body-name: data
    delete:
      tags:
        - Shop
      operationId: shop_delete
      parameters:
        - name: id
          in: path
          description: A unique integer value identifying this shop.
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: ''
          content: {}
  /shop/{shop}/services:
    get:
      tags:
        - Shop
      description: Returns a list of all services provided by a shop
      operationId: shop_services_list
      parameters:
        - name: shop
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content: {}
    put:
      tags:
        - Shop
      description: Update a given shop service
      operationId: shop_services_update
      parameters:
        - name: shop
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopService'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopService'
      x-codegen-request-body-name: data
    post:
      tags:
        - Shop
      description: Create a new service for a given shop
      operationId: shop_services_create
      parameters:
        - name: shop
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopService'
        required: true
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopService'
      x-codegen-request-body-name: data
    delete:
      tags:
        - Shop
      description: Returns a list of all services provided by a shop
      operationId: shop_services_delete
      parameters:
        - name: shop
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                service_id:
                  type: string
                  description: int
        required: true
      responses:
        '204':
          description: ''
          content: {}
      x-codegen-request-body-name: data
...
components:
  schemas:
...
    Shop:
      required:
        - email
        - name
        - owner
      type: object
      properties:
        id:
          title: ID
          type: integer
          readOnly: true
        name:
          title: Name
          maxLength: 255
          minLength: 1
          type: string
        address:
          title: Address
          maxLength: 500
          type: string
        description:
          title: Description
          maxLength: 500
          type: string
        email:
          title: Email
          maxLength: 75
          minLength: 1
          type: string
          format: email
        phone:
          title: Phone
          maxLength: 15
          minLength: 1
          pattern: ^\d{3}-\d{3}-\d{4}$
          type: string
          nullable: true
        hours:
          title: Hours
          maxLength: 255
          minLength: 1
          type: string
          nullable: true
        latitude:
          title: Latitude
          type: string
          format: decimal
          nullable: true
        longitude:
          title: Longitude
          type: string
          format: decimal
          nullable: true
        owner:
          title: Owner
          type: string
        services:
          uniqueItems: true
          type: array
          items:
            type: integer
        employees:
          uniqueItems: true
          type: array
          items:
            type: string
...
    ShopService:
      required:
        - service
      type: object
      properties:
        id:
          title: ID
          type: integer
          readOnly: true
        service:
          title: Service
          type: integer
        price:
          title: Price
          type: string
          format: decimal
          nullable: true
        est_time:
          title: Est time
          type: string
...
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header